Refactors API import paths
Updates import paths for API-related files to improve code organization and maintainability.

This change centralizes API definitions under a common directory and uses shared types, enhancing code clarity and reducing potential import errors across the application.
