+++ src/pages/ServerSettings.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import { FaSave, FaArrowLeft } from "react-icons/fa";
-import { fetchServerSettings, updateServerSettings } from "../utils/api";
+import { fetchServerSettings, updateServerSettings } from "../utils/api/OpenVpnServerCerts";
 import { toast } from "react-toastify";
 
 import "../css/ServerSettings.css";
