+++ src/pages/ServersOverview/helpers.ts
@@ -3,7 +3,7 @@ import type { ChartPoint } from "./types";
 import type {
   OverviewSeriesResponse,
   OverviewSeriesRow,
-} from "../../utils/api";
+} from "../../utils/types";
 
 /* ---- time helpers ---- */
 export function startOfToday() { const n = new Date(); n.setHours(0,0,0,0); return n; }
